Define basic_indices inside NumPy extra
Prior factory pattern obfuscates method signature
